FibClock works!
This is a gross first pass. I do a huge enumeration of all possible
hour/minute combinations (couldn't think of a better way to do this
in Atom). Right now I am just taking the first possibility of the
enumerated solutions, but I would like to do a random choice like
the KickStarter project does.

This produces a 182K .ino file which is absurdly huge. Oh well,
I get to do all the "heavy lifting" in Haskell.